product done

+++ src/main/java/store/product/ProductIn.java
@@ -5,6 +5,6 @@ import lombok.Builder;
 @Builder
 public record ProductIn(
     String name,
-    Double price,
+    Float price,
     String unit
 ) {}
